cleanup code, wrap svg in button
